+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Allow `instance_uri` configuration value to contain URI scheme.
 - Changed `Content-Security-Policy` header value in nginx config examples.
+- Changed `/api/v1/{status_id}/context` response format to match Mastodon API.
 
 ## [1.13.1] - 2023-02-09